% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_chooser.R
\name{game_chooser}
\alias{game_chooser}
\title{Shiny app to help choose a game to play}
\usage{
game_chooser(...)
}
\arguments{
\item{...}{options passed to \code{\link[shiny:shinyApp]{shiny::shinyApp()}}}
}
\description{
This app uses a board game collection (as accessed from
\href{https://boardgamegeek.com}{boardgamegeek.com}) to help a user choose a game
to play based on the number of players, the time available, and/or the age
of the youngest player. There's also a random game selector if you're
really indecisive!
}
