% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_surround.R
\name{get_cols}
\alias{get_cols}
\title{Get all column vectors of a matrix.}
\usage{
get_cols(data)
}
\arguments{
\item{data}{Matrix from which to extract column vectors.}
}
\description{
This function extracts all column vectors of a matrix and returns the result as a list.
}
\examples{
M = matrix(rnorm(9),3,3)
get_cols(M)
}
\keyword{column}
\keyword{extract}
\keyword{matrix}
\keyword{vectors}

