% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_surround.R
\name{get_diags}
\alias{get_diags}
\title{Get all diagonals vectors of a matrix.}
\usage{
get_diags(data, direction = "right")
}
\arguments{
\item{data}{Matrix from which to extract diagonal elements}

\item{direction}{Which side to begin on? Takes values of one of "left", "right" or "both". Defaults to "right".}
}
\description{
This function extracts all diagonal vectors of a matrix and returns the result as a list.
}
\examples{
M = matrix(rnorm(9),3,3)
get_diags(M)
}
\keyword{diagonal}
\keyword{matrix}
\keyword{vectors}

