% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_surround.R
\name{get_rows}
\alias{get_rows}
\title{Get all row vectors of a matrix.}
\usage{
get_rows(data)
}
\arguments{
\item{data}{Matrix from which to extract row vectors.}
}
\description{
This function extracts all row vectors of a matrix and returns the result as a list.
}
\examples{
M = matrix(rnorm(9),3,3)
get_rows(M)
}
\keyword{extract}
\keyword{matrix}
\keyword{row}
\keyword{vectors}

