% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy2index.R
\name{index2xy}
\alias{index2xy}
\title{Converts a matrix index into a sex of x,y coordinates.}
\usage{
index2xy(data, index)
}
\arguments{
\item{data}{Matrix or data frame.}

\item{index}{A vector of index values.}
}
\description{
This function converts a matrix index into unit x,y plotting coordinates.
}
\examples{
M = matrix(1:20,4,5)
index2xy(data = M, index = c(3,4))
}
\keyword{convert}
\keyword{coordinate}
\keyword{element}
\keyword{index}
\keyword{matrix}

