% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy2index.R
\name{is_palindrome}
\alias{is_palindrome}
\title{Palindrome checker.}
\usage{
is_palindrome(x, case.sensitive = FALSE)
}
\arguments{
\item{x}{Numeric or character vector.}

\item{case.sensitive}{Does upper or lower casing matter? Defaults to FALSE.}
}
\description{
This function checks if the supplied vector is a palindrome (reads the same forwards and backwards).
}
\examples{
test1 = 123
test2 = "12321"
test3 = c("a",1,2,3,2,1,"a")
is_palindrome(test1)
is_palindrome(test2)
is_palindrome(test3)
}
\keyword{case}
\keyword{check}
\keyword{palindrome}
\keyword{sensitive}
\keyword{vector}

