% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TicTacToe.R
\name{UltimateTicTacToe}
\alias{UltimateTicTacToe}
\title{Play some Ultimate Tic-Tac-Toe?}
\usage{
UltimateTicTacToe()
}
\description{
This function allows one to play the Ultimate version of Tic-Tac-Toe.
In the Regular version of Tic-Tac-Toe, players take turns placing their marks, with the objective of achieving three marks in a row in any direction.
9x9 Tic-Tac-Toe or more commonly known as Ultimate Tic-Tac-Toe, adds a twist on the regular version of
Tic-Tac-Toe that most of us have come to know. Perceive the board as a big Tic-Tac-Toe board, with
the goal being to achieve 3 big marks in any direction. Big marks are achieved by winning the corresponding
small Tic-Tac-Toe blocks. The player to move first may play anywhere on the board. However, following moves
must correspond to the same big Tic-Tac-Toe block of the small Tic-Tac-Toe board where the last move was played.
}
\keyword{connect}
\keyword{dataframe}
\keyword{fun}
\keyword{game}
\keyword{matrix}
\keyword{row}
\keyword{tac}
\keyword{tic}
\keyword{tictactoe}
\keyword{toe}
\keyword{ultimate}

